Finished structure of landing (main) page. Restructured file directory.
