--- frontend/formula-hacks-app/src/FoodList.css
@@ -1,18 +1,0 @@
-.food-list-container {
-    display: flex;
-    justify-content: center;
-    align-items: center;
-    height: 100vh; /* Adjust this value if needed */
-  }
-  
-  .food-list {
-    text-align: center;
-  }
-
-  .fetch-button {
-    margin-top: 20px;
-  }
-  
-  .no-meal-available {
-    margin-top: 20px;
-  }
