+++ frontend/src/Profile.css
@@ -1,0 +1,13 @@
+/* Profile.css */
+
+.food-cards-container {
+    display: flex;
+    flex-wrap: nowrap;
+    overflow-x: auto; /* Enable horizontal scrolling */
+    padding: 10px; /* Add padding to the container */
+  }
+  
+  .favorite-card {
+    margin-right: 10px; /* Add margin to the right of each card */
+  }
+
