+++ frontend/src/assets/profile-picture.png
