+++ frontend/src/assets/spaghetti.jpg
